I have refactored the codebase to improve modularity and readability.
I separated concerns by creating modules for file parsing, distance calculations, ordination calculations, and plotting. This makes the code easier to maintain and understand.
