+++ index.html
@@ -41,6 +41,6 @@
     <div id="plot"></div>
   </div>
 
-  <script src="script.js"></script>
+  <script type="module" src="script.js"></script>
 </body>
 </html>
